requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai
langchain-community

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Lang Graph
langgraph

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

matplotlib
pandas
streamlit